requirements.txt

# CameraHMR official reqs
git+https://github.com/facebookresearch/detectron2.git
pytorch-lightning==2.0.2
smplx==0.1.28
pyrender==0.1.45
opencv-python==4.10.0.84
yacs==0.1.8
scikit-image==0.21.0
einops==0.8.0
timm==1.0.12
pandas==2.1.1
loguru==0.7.3
numpy==1.23.5
pyrootutils==1.0.4
rich==14.0.0
hydra-colorlog==1.2.0
albumentations==1.3.1
chumpy
flatten_dict
jpeg4py
joblib

# IBRICS reqs
ffmpeg-python
